# drs_server/drs_openapi.yaml
openapi: 3.0.3
info:
  title: Data Repository Service
  version: 1.3.0
  termsOfService: 'https://www.ga4gh.org/terms-and-conditions/'
  license:
    name: Apache 2.0
    url: 'https://raw.githubusercontent.com/ga4gh/data-repository-service-schemas/master/LICENSE'
servers:
  - url: /ga4gh/drs/v1
security:
  - {}
  - BasicAuth: []
  - BearerAuth: []
paths:
...
  /objects:
    parameters:
      - $ref: '#/components/parameters/BearerAuth'
    get:
      summary: List all DRS objects
      parameters:
      - $ref: '#/components/parameters/ProgramIdQuery'
      - $ref: '#/components/parameters/SampleRegistrationQuery'
      description: >-
        Returns object metadata, and a list of access methods that can be used to fetch object bytes.
      operationId: drs_operations.list_objects
      responses:
        200:
          description: The `DrsObject` was found successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  anyOf:
                    - $ref: '#/components/schemas/ExperimentDrsObject'
                    - $ref: '#/components/schemas/AnalysisDrsObject'
                    - $ref: '#/components/schemas/DataDrsObject'
      tags:
        - Objects
    post:
        summary: Add a DrsObject to the db
        operationId: drs_operations.post_object
        requestBody:
            $ref: '#/components/requestBodies/DrsObjectRequest'
        responses:
            200:
              description: The `DrsObject` was found successfully
              content:
                application/json:
                  schema:
                    type: object
  /objects/{object_id}:
    parameters:
        - $ref: '#/components/parameters/ObjectId'
        - $ref: '#/components/parameters/BearerAuth'
    delete:
      summary: Delete a DrsObject.
      description: >-
        Returns object metadata, and a list of access methods that can be used to fetch object bytes.
      operationId: drs_operations.delete_object
      responses:
          200:
            description: The `DrsObject` was found successfully
            content:
              application/json:
                schema:
                  type: object
    get:
      summary: Get info about a DrsObject.
      description: >-
        Returns object metadata, and a list of access methods that can be used to fetch object bytes.
      operationId: drs_operations.get_object
      parameters:
        - $ref: '#/components/parameters/Expand'
      responses:
        200:
          description: The `DrsObject` was found successfully
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/AnalysisDrsObject'
      tags:
        - Objects
  /objects/{object_id}/download:
    parameters:
        - $ref: '#/components/parameters/ObjectId'
        - $ref: '#/components/parameters/BearerAuth'
    get:
      summary: Download a file associated with a drs object
      description: downloads the file.
      operationId: drs_operations.download_file
      responses:
        200:
          description: The `DrsObject` was found successfully
          content:
            application/octet-stream: {}
      tags:
        - Objects
  /objects/{object_id}/access/{access_id}:
    get:
      summary: Get a URL for fetching bytes
      description: >-
        Returns a URL that can be used to fetch the bytes of a `DrsObject`.
      operationId: drs_operations.get_access_url
      parameters:
        - $ref: '#/components/parameters/BearerAuth'
        - $ref: '#/components/parameters/ObjectId'
        - $ref: '#/components/parameters/AccessId'
      responses:
        200:
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                    description: full URL required to get bytes
                  headers:
                    type: array
                    items:
                      type: string
                    description: array of strings that need to be passed with the URL to get bytes
      tags:
        - Objects
  /biosamples:
    parameters:
     - $ref: '#/components/parameters/BearerAuth'
    post:
      summary: List all DRS Experiment objects matching an array of submitter_sample_ids. Only authorized for site admins and authzed services.
      operationId: drs_operations.list_biosamples
      requestBody:
        $ref: '#/components/requestBodies/BiosamplesRequest'
      responses:
        200:
          description: The `DrsObjects` were found successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  anyOf:
                    - $ref: '#/components/schemas/BiosampleTicket'
      tags:
        - Objects
  /programs:
    parameters:
      - $ref: '#/components/parameters/BearerAuth'
    get:
      description: List all programs available
      operationId: drs_operations.list_programs
      responses:
        200:
          description: ok
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                description: array of programs
    post:
      description: Create a program
      operationId: drs_operations.post_program
      requestBody:
          $ref: '#/components/requestBodies/ProgramRequest'
      responses:
          200:
            description: The `Program` was created successfully
            content:
              application/json:
                  schema:
                    type: object
  /programs/{program_id}:
    parameters:
      - $ref: "#/components/parameters/ProgramId"
      - $ref: '#/components/parameters/BearerAuth'
    get:
      description: Describe a program
      operationId: drs_operations.get_program
      responses:
        200:
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Program'
    delete:
      summary: Delete a program.
      description: >-
        Delete a program, and all associated DRS objects.
      operationId: drs_operations.delete_program
      responses:
          200:
            description: The program was deleted successfully
            content:
              application/json:
                schema:
                  type: object
  /programs/{program_id}/status:
    parameters:
      - $ref: "#/components/parameters/ProgramId"
      - $ref: '#/components/parameters/BearerAuth'
    get:
      description: List indexing status reports for a program
      operationId: drs_operations.get_program_status
      responses:
        200:
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProgramIndexStatus'
components:
  parameters:
    BearerAuth:
      in: header
      name: Authorization
      schema:
        type: string
        description: A string of the format 'Bearer <token>'
    ProgramId:
      in: path
      name: program_id
      required: true
      description: ProgramId
      schema:
        type: string
    ProgramIdQuery:
        in: query
        name: program_id
        description: ProgramId
        schema:
          type: string
    SampleRegistrationQuery:
        in: query
        name: submitter_sample_id
        description: submitter_sample_id from mohccn data model
        schema:
          type: string
    ObjectId:
      in: path
      name: object_id
      required: true
      description: '`DrsObject` identifier'
      schema:
        type: string
        format: path
    Expand:
      in: query
      name: expand
      schema:
        type: boolean
      example: false
      description: >-
        If false and the object_id refers to a bundle, then the ContentsObject array
        contains only those objects directly contained in the bundle. That is, if the
        bundle contains other bundles, those other bundles are not recursively
        included in the result.
        If true and the object_id refers to a bundle, then the entire set of objects
        in the bundle is expanded. That is, if the bundle contains aother bundles,
        then those other bundles are recursively expanded and included in the result.
        Recursion continues through the entire sub-tree of the bundle.
        If the object_id refers to a blob, then the query parameter is ignored.
    AccessId:
      in: path
      name: access_id
      required: true
      description: An `access_id` from the `access_methods` list of a `DrsObject`
      schema:
        type: string
        format: path
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
      description: |
        A valid authorization token must be passed in the 'Authorization' header,
        e.g. "Basic ${token_string}"
    BearerAuth:
      type: http
      scheme: bearer
      description:
        A valid authorization token must be passed in the 'Authorization' header,
        e.g. "Bearer ${token_string}"
  requestBodies:
    DrsObjectRequest:
      content:
        'application/json':
          schema:
            anyOf:
              - $ref: "#/components/schemas/ExperimentDrsObject"
              - $ref: "#/components/schemas/AnalysisDrsObject"
              - $ref: "#/components/schemas/DataDrsObject"
    ProgramRequest:
      content:
        'application/json':
          schema:
            $ref: '#/components/schemas/Program'
    BiosamplesRequest:
      description: list of submitter sample ids requested
      required: false
      content:
        'application/json':
          schema:
            type: object
            properties:
              program_ids:
                type: array
                items:
                  type: string
              submitter_sample_ids:
                type: array
                items:
                  type: string
  schemas:
    Program:
      type: object
      properties:
        id:
          type: string
          description: name of program
        drsobjects:
          type: array
          items:
            $ref: '#/components/schemas/DrsUri'
        statistics:
          type: object
    ProgramIndexStatus:
      type: object
      properties:
        index_complete:
          description: DrsUris for variant objects that have been successfully indexed
          type: array
          items:
            $ref: '#/components/schemas/DrsUri'
        index_in_progress:
          description: DrsUris for variant objects that have not yet been indexed
          type: array
          items:
            $ref: '#/components/schemas/DrsUri'
        index_errored:
          description: DrsUris for variant objects that are in the queue to be indexed
          type: array
          items:
            type: object
            properties:
              drs_uri:
                $ref: '#/components/schemas/DrsUri'
              errors:
                type: array
                items:
                  type: string
    DrsUri:
      type: string
      description: a DRS object's self_uri
    BiosampleTicket:
      type: object
      properties:
        biosample_id:
          type: string
          description: MoH submitter_sample_id from a Sample Registration associated with a specimen.
        program:
          type: string
          description: MoH program ID
        experiments:
          type: array
          description: An array of associated ExperimentDrsObjects derived from the biosample.
          items:
            type: string
        runs:
          type: array
          description: An array of associated RunDrsObjects that describe raw reads derived from biosample.
          items:
            type: string
        analyses:
          type: object
          description: Arrays of associated AnalysisDrsObjects based on the biosample, keyed by analysis type.
          properties:
            reference_alignment:
              type: array
              items:
                type: string
            sequence_variation:
              type: array
              items:
                type: string
            sequence_annotation:
              type: array
              items:
                type: string
    ExperimentDrsObject:
      type: object
      description: A DrsObject that describes the sequencing experiment.
      required:
        - name
        - id
        - contents
        - description
        - program
      properties:
        name:
          type: string
          description: The biosample identifier for the experiment.
        id:
          type: string
          description: A unique ID for the experiment. Often the alias for the experiment, as defined by the sequencing center
        contents:
          type: array
          description: The specific analysis contents objects that were generated from this experiment.
          items:
            $ref: '#/components/schemas/ContentsObject'
        description:
          type: string
          description: the library strategy used in this experiment
          enum:
            - wgs
            - wts
        program:
          type: string
          description: The program this object was ingested as part of
        metadata:
          type: object
          description: metadata about the experiment (see Experiment in ingest_openapi.yaml)
        aliases:
          type: array
          items:
            type: string
          description: >-
            A list of strings that can be used to find other metadata
            about this `DrsObject` from external metadata sources. These
            aliases can be used to represent secondary
            accession numbers or external GUIDs.
        self_uri:
          type: string
          description: |-
            A drs:// hostname-based URI, as defined in the DRS documentation, that tells clients how to access this object.
            The intent of this field is to make DRS objects self-contained, and therefore easier for clients to store and pass around.  For example, if you arrive at this DRS JSON by resolving a compact identifier-based DRS URI, the `self_uri` presents you with a hostname and properly encoded DRS ID for use in subsequent `access` endpoint calls.
          example:
            drs://drs.example.org/314159
        size:
          type: integer
          format: int64
          description: The cumulative size, in bytes, of items in the `contents` field.
        created_time:
          type: string
          format: date-time
          description: |-
            Timestamp of content creation in RFC3339.
            (This is the creation time of the underlying content, not of the JSON object.)
        updated_time:
          type: string
          format: date-time
          description: >-
            Timestamp of content update in RFC3339, identical to `created_time` in systems
            that do not support updates.
            (This is the update time of the underlying content, not of the JSON object.)
        version:
          type: string
          description: >-
            A string representing a version.
            (Some systems may use checksum, a RFC3339 timestamp, or an incrementing version number.)
        checksums:
          type: array
          # minItems: 1
          items:
            $ref: '#/components/schemas/Checksum'
    AnalysisDrsObject:
      type: object
      description: A DrsObject that describes a sequencing analysis. It usually will consist of a data file, e.g. a variant or read file, and its associated index file. Its contents should also include any associated Experiments (as ContentObjects), ordered in order of appearance in the associated variant/read files.
      required:
        - name
        - id
        - contents
        - description
        - program
        - reference_genome
      properties:
        name:
          type: string
          description: The unique ID for this analysis
        id:
          type: string
          description: same as name
        contents:
          type: array
          description: The specific analysis contents objects that comprise this analysis DRS entity. Should contain one or more ContentsObjects that describe the data files in the analysis, and one or more ContentsObjects corresponding to the experiments analyzed in the analysis.
          minItems: 1
          items:
            $ref: '#/components/schemas/ContentsObject'
        description:
          type: string
          description: type of sequencing analysis
          enum:
            - reference_alignment
            - sequence_variation
            - sequence_annotation
        program:
          type: string
          description: The program this object was ingested as part of
        reference_genome:
          type: string
          description: the reference genome used to compile this analysis
          enum:
            - hg37
            - hg38
        indexed:
          type: integer
          description: 1 if the object has been indexed for search, 0 if not
        aliases:
          type: array
          items:
            type: string
          description: >-
            A list of strings that can be used to find other metadata
            about this `DrsObject` from external metadata sources. These
            aliases can be used to represent secondary
            accession numbers or external GUIDs.
        self_uri:
          type: string
          description: |-
            A drs:// hostname-based URI, as defined in the DRS documentation, that tells clients how to access this object.
            The intent of this field is to make DRS objects self-contained, and therefore easier for clients to store and pass around.  For example, if you arrive at this DRS JSON by resolving a compact identifier-based DRS URI, the `self_uri` presents you with a hostname and properly encoded DRS ID for use in subsequent `access` endpoint calls.
          example:
            drs://drs.example.org/314159
        size:
          type: integer
          format: int64
          description: The cumulative size, in bytes, of items in the `contents` field.
        created_time:
          type: string
          format: date-time
          description: |-
            Timestamp of content creation in RFC3339.
            (This is the creation time of the underlying content, not of the JSON object.)
        updated_time:
          type: string
          format: date-time
          description: >-
            Timestamp of content update in RFC3339, identical to `created_time` in systems
            that do not support updates.
            (This is the update time of the underlying content, not of the JSON object.)
        version:
          type: string
          description: >-
            A string representing a version.
            (Some systems may use checksum, a RFC3339 timestamp, or an incrementing version number.)
        checksums:
          type: array
          # minItems: 1
          items:
            $ref: '#/components/schemas/Checksum'
        metadata:
          type: object
          description: metadata about the analysis (see Analysis in ingest_openapi.yaml)
    DataDrsObject:
      type: object
      description: A DrsObject that describes a data file, e.g. a variant or read file.
      required:
        - name
        - id
        - description
        - access_methods
        - program
      properties:
        name:
          type: string
          description: Unique ID for this data object
        id:
          type: string
          description: The filename for this data file, including its file extensions.
        description:
          type: string
          description: type of file
        access_methods:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/AccessMethod"
          description: |-
            The list of access methods that can be used to fetch the `DrsObject`.
        program:
            type: string
            description: The program this object was ingested as part of
    Checksum:
      type: object
      required:
        - checksum
        - type
      properties:
        checksum:
          type: string
          description: The hex-string encoded checksum for the data
        type:
          type: string
          description: >-
            The digest method used to create the checksum.
            The value (e.g. `sha-256`) SHOULD be listed as `Hash Name String` in the https://www.iana.org/assignments/named-information/named-information.xhtml#hash-alg[IANA Named Information Hash Algorithm Registry].
            Other values MAY be used, as long as implementors are aware of the issues discussed in https://tools.ietf.org/html/rfc6920#section-9.4[RFC6920].
            GA4GH may provide more explicit guidance for use of non-IANA-registered algorithms in the future.
            Until then, if implementors do choose such an algorithm (e.g. because it's implemented by their storage provider), they SHOULD use an existing
            standard `type` value such as `md5`, `etag`, `crc32c`, `trunc512`, or `sha1`.
          example: sha-256
...
    AccessMethod:
      anyOf:
        - $ref: '#/components/schemas/AccessUrlMethod'
        - $ref: '#/components/schemas/AccessIdMethod'
    ContentsObject:
      type: object
      required:
        - name
        - id
        - drs_uri
      properties:
        name:
          type: string
          description: The identifier of the linked DRS object
        id:
          type: string
          description: Alternate identifier for the linked DRS object
        drs_uri:
          type: array
          description: The DRS uri(s) to the linked DrsObject
          items:
            type: string